Split pagination code
